Add back project_environments validation for pre 7.53.1
